refactor: Split insert.rs into focused submodules
Refactored the 504-line executor/src/insert.rs into a modular structure
with clear separation of concerns:

- insert/mod.rs (21 lines) - Main InsertExecutor and public API
- insert/validation.rs (172 lines) - Input validation and type coercion
- insert/constraints.rs (155 lines) - NOT NULL, PRIMARY KEY, UNIQUE, CHECK constraints
- insert/foreign_keys.rs (47 lines) - Foreign key constraint checking
- insert/defaults.rs (82 lines) - Default value handling and generation
- insert/execution.rs (147 lines) - Orchestrates validation and insertion

Benefits:
- Each module has a single, well-defined responsibility
- Easier to navigate and understand INSERT implementation
- Constraint checking logic can be shared with UPDATE executor
- All 297 executor unit tests pass
- No functional changes or performance degradation
